DM-118 Organized sleeper search files into a folder
