+++ frontend/src/feature/search/sleeper/hooks/useDeleteLeague.ts
